Authenticate API calls with HTTP Basic auth.
A request is considered to be an API call if the HTTP ACCEPT header prefers application/json.

If an API call is detected then:
  * the omniauth strategy will not be run at all
  * a warden strategy :gds_sso_api_access or :mock_gds_sso_api_access will be run.
    This is controlled by the #valid? method on the strategy.

The mock api_access strategy always authenticates the first user irrespective of the
supplied credentials.

The production strategy will look up the credentials in the config. To configure set:

GDS::SSO.config do |config|
  config.basic_auth_user = 'api'
  config.basic_auth_password = 'secret'
end