Add gds_bearer_token warden strategy
Providing an OAuth2 bearer token (using the Authorization header) will now cause an
app using gds-sso to look that user up in signonotron and authorize them accordingly.
This means that we can treat API clients as users and grant them permissions just
as we would anyone/anything else.

This should be treated as a work in progress as one test is currently failing, with
the containing app returning a 500 rather than a 401 if an authorization attempt
fails.

It's also worth noting that for the time being the HTTP Basic mechanism remains in
place. That's to allow for backwards compatibility while we migrate all existing API
clients to the new mechanism.