Support permissions from Signon as Array or Hash
This enables us to gradually upgrade each app to support new and old
formats, then when we're ready, we can change Signon to emit Arrays.

The reason for the format change is because it's unneeded complexity
that existed when we provided the permissions for a user for every app
to every app. That is bad for security in terms of info sharing: it
tells you the valuable user accounts.