Fix failing test with Devise 2.2.x
Devise 2.2.x changed behaviour when it thinks the request isn't
"navigational" (ie not a browser requesting HTML) - it will return a 204
instead of a redirect which is what it does if it thinks you want a
navigational response.

I couldn't work out if there's a better fix (some additional headers on the
request maybe?), and this test would fail if the sign out hadn't worked anyway.