make dependency on omniauth-gds less restrictive
I'd like to make a minor change to omniauth-gds (publish it to rubygems
instead of gemfury) and because of the exact dependency that gds-sso has
on omniauth-gds, this means that I have to release a new version of
gds-sso to pick up this change.

Using a ~> dependency would be a bit more lenient in future.