Including GDS::SSO::ControllerMethods into auth controller
We had an odd issue where we only included GDS::SSO::ControllerMethods
in our admin controller, however as AuthenticationsController inherits
from ApplicationController it needs to be included here instead, so
that authenticate_user! is defined.

Alternatively the readme should be updated to state that
GDS::SSO::ControllerMethods has to be included in the
ApplicationController