Use MySQL for running Signon
This fixes an error we are seeing when running the integration tests since
upgrading Signon to Rails 4.0.x:

  ArgumentError: Index name 'index_batch_invite_app_perms_on_batch_invite_and_supported_perm' on table 'batch_invitation_application_permissions' is too long; the limit is 62 characters

Signon's index names are limited to 64 characters as this is the limit for
MySQL (in Rails). For sqlite, Rails' limit is 62 characters. See:
https://github.com/rails/rails/blob/v4.0.13/activerecord/lib/active_record/connection_adapters/sqlite3_adapter.rb#L193-L198

Signon is developed and tested against MySQL, so let's stop expecting it to
work on sqlite.