Add a script for building against signon branches
When reviewing a signon pull request, we want to know if gds-sso is compatible
with the changes. This script will be used by a new CI build and will record
the status of the gds-sso build against the signon commit and GitHub will
display the multiple statuses, just as we do for govuk-content-schemas PRs.