Add password_changed_at to fixture data
With the updates to the password_expiry module in signonotron, it now
requires a password change if this field is `NULL`. In practice, this
field is set by ActiveRecord on creation, and in production data only
one user has `NULL`, which is an API user which is unaffected by
password expiry.

This fixes the integration tests that run against signonotron by
explicitly specifying that the user’s password is “fresh”.