Remove WWTD gem and explicitly define testing
This repo made use of the "wwtd" gem (what would travis do) in order to
run a pseudo build matrix based on gemfile definitions in `travis.yml`.
The gem is fairly opaque and runs `bundle install` with `--quiet` by
default, with no way of overriding it. This was masking errors in CI
which, it turned out, were caused by the gem anyway.

This PR removes wwtd and unpacks its behaviour into a couple of bash
loops, making the behaviour more explicit and obvious. And fixes the
error.

The `signonotron:stop` task in the bash trap directive has been
replaced with a pure shell kill command, due to the uncertain nature of
the bundle environment at the time the trap is called.