Pin rack-cache to 1.2 if building for Ruby 1.9
rack-cache 1.3.0 has been released which removes support for Ruby 1.9.
`gds-sso` itself still supports Ruby 1.9, this dependency comes from
`actionpack` and is pessimistically specified as `~> 1.2`, which
permits `1.3.0`

This change ensures that if we’re building the gem with a fresh
lockfile under Ruby 1.9, we pin rack-cache to a compatible version.
This will not affect behaviour of the gem.