Add `User` linting for `ActiveSupport::TestCase`
While this project had a useful linter compatible with RSpec, it was
sadly missing a regular test-unit and minitest compatible equivalent.

Given the absence of an RSpec-like `described_class` and
`shared_examples` I invented something similar using `user_class` and
regular reopening. Usage is detailed in the `UserTest` header comment.