Bump `capybara` and `capybara-mechanize`
We were using a very old vintage of capybara, that is no longer
compatible with Rack 2. Also, newer versions of capybara are
unsurprisingly not compatible with the earlier versions of
capybara-mechanize, so that was bumped too.

The Gemfile was explicitly pinning mechanize, which now seems
redundant. We can allow that to resolve transitively.

In bumping capybara from 1.x to 2.x we hit an issue where we were
depending on broken behaviour [0] to exercise a particular vector. I've
marked these specs pending for now as I couldn't figure a way to ensure
and assert that the correct authentication dance had been followed (at
least without hacks into the driver). Other perhaps semi-related
behaviour that may shed some light [1][2].

I verified by hand that the original behaviour is preserved, but
obviously it would be beneficial for us to resolve these upstream, or
figure a way to test this behaviour effectively.

[0] https://github.com/jnicklas/capybara/issues/1023
[1] https://github.com/jnicklas/capybara/pull/603
[2] https://github.com/jnicklas/capybara/pull/606