Fix `render :text` deprecations
Rails 5 was spewing:

```
DEPRECATION WARNING: `render :text` is deprecated because it does not
actually render a `text/plain` response. Switch to `render plain: 'plain
text'` to render as `text/plain`, `render html: '<strong>HTML</stron
g>'` to render as `text/html`, or `render body: 'raw'` to match the
deprecated behavior and render with the default Content-Type, which is
`text/plain`.
```

Replacing our use of `text` with `body` resolves this.