Don't rely on capybara referrer for spec
The referrer is no longer preserved in capybara when redirects happen.
We were relying on this broken behaviour to test that an expired app
session would re-check with signonotron. Instead, we now force a logged
out state in signon if appropriate, and check that we’re either allowed
through or asked to sign in again.