Update signon fixture password
In alphagov/signonotron2@bceee5dd7208c00697eea801355ab5efd4c2ff0d
I changed the test environment to start using a pepper, which more
closely mimics production.

This means that the encrypted password in this repo is no longer
valid, so I've created a new user at the console with the same
password and updated the encrypted values here.