Deprecate require_signin_permission! controller method
Changes to the signon application to implement
[RFC-78](https://github.com/alphagov/govuk-rfcs/blob/8cbb2a0de86de02f54ae37b245e79b46ad62cb6a/rfc-078-re-architect-signin-permissions-in-signon.md)
mean it signin permission is no longer optional for users accessing an
application via oauth.  Signon itself will make sure the signin permission
has been granted to the user and reject the oauth handshake if they do not
have it.  We deprecate the `require_signin_permission!` method and update
the docs to make this clear.