Return JSON for auth failures in api_only
This stops the behaviour being to redirect someone to signon when
they're trying to access a resource without a bearer token and this gem
is in api_only mode.

It also renders JSON responses and to do that with
ActionController::Metal you need to import a few modules as it's pretty
bare bones.