Tweak the .gitignore file
Remove the Gemfile.lock, this file is committed to most repositories,
and I think doing so would help avoid dependency resolution issues.

Also remove .DS_Store, as that's something to do with macOS (as far as
I know), so it should be included in more general .gitignore
files (e.g. the user one).