Set a lower default connection timeout for connecting to Signon
I believe Faraday is using the Net::HTTP adapter by default, and the
default open_timeout value there is 60 seconds. Connections to Signon
should be much quicker than that, so shorter the default open_timeout.

This should help avoid requests queuing up if Signon is temporarily
unavailable, and this change is motivated by a recent issue with this
happening to the Content Store.