Modernise ruby version and what we test
This tests this gem against Ruby 2.6 and 2.7 (since 2.6 is the minimum
GOV.UK use), and marks that as the minimum supported version in the
gemspec. It also switches the .ruby-version of this repo to be 2.6.6 to
reflect GOV.UK doesn't use 2.5 in production anymore.

This also updates the Rails versions tested to be Rails 5 and 6 as these
are the versions currently used at GOV.UK.

It removes the redundant omniauth-oauth2 from the Rails gemfiles since
this is a dependency of omniauth-gds in the gemspec.