DRY up config for consumers of this gem
Previously we required each consumer of this gem to have similar
initializers, with the following code:

    GDS::SSO.config do |config|
      config.user_model = "User"
      config.oauth_id = ENV.fetch("OAUTH_ID", "abcdefg")
      config.oauth_secret = ENV.fetch("OAUTH_SECRET", "secret")
      config.oauth_root_url = Plek.new.external_url_for("signon")*
    end

This updates the default config to match the common case for our
initializers, so that a given app only needs to specify additional
config, such as Rails caching.

Making this change removes a body of untested code that we would
otherwise need to consider testing before we could enable automatic
deployment for an app.

*Note that some apps use "Plek.find", which returns the internal
hostname for Signon. Since the external domain works for all apps,
this will switch all of them to use it.