Prefix environment variables with GDS_SSO
This prefix makes it easier to understand that these environment variables
are for this particular gem. Currently a GOV.UK app has the OAUTH_ID and
OAUTH_SECRET environment variables but they don't tend to appear in an
apps code as they're used automatically by this gem. By making this
change it makes it clearer they have a purpose for a particular tool.

This approach is consistent with a number of other GOV.UK gems for
example: https://github.com/alphagov/plek#environment-variables and
https://github.com/alphagov/govuk_test#usage