Allow OmniAuth to accept GET requests
During initial authentication the FailureApp will redirect users to
"/auth/gds".

As part of remediating CVE 2015 9284, OmniAuth v2.0 no longer allows
plain GET requests.

https://github.com/omniauth/omniauth/wiki/Resolving-CVE-2015-9284

A full remediation of this vulnerability will involve introducing an
intermediate step to turn the automatic GET
redirect into a manual POST request. This process could be automated
using JS to submit the form.

Until this can be implemented, and to allow us to continue to
upgrade OmniAuth, it is possible to allow GET requests with the
caveat that this does not resolve the vulnerabilty.

This future work has been recorded in this Trello ticket
https://trello.com/c/eSW7oKde/2342-use-post-requests-during-authentication-in-gds-sso