Remove the end-to-end signon tests and supplementary files
Following from the last commit which introduced replacement tests, this
removes the signon integration tests.

These were unusual tests that required a running instance of the signon
application. These were a complicated coupling that required this gem to
know a lot about how to set-up signon (including db schema) whcih caused
a tight coupling. They are difficult and slow to run, and a source of
surprise for anyone working with this gem.

These tests were only run when gds-sso changed and not when signon
changed, so any assurances they made about integration would not be
guaranteed if signon changed later.