Create a system test for testing a GDS SSO Rails integration
This is the introduction of RSpec system tests that are intended to
replace the spec/requests/end_to_end_spec.rb which relies on a running
instance of Signon.

The motivation to create these new tests is that running the tests that
rely on a running instance of signon is both difficult and slow (both in
setup and execution). The cost of the extra effort to deal with these
exceeds any benefits we have it being a more thorough suite of tests.

These tests aim to match the scenarios we were previously trying to
cover in the end to end test, but this tries to be more focused on
explicit scenarios.

To set these tests up I had to turn off Capybara redirect following so
that Capybara wouldn't try to request a fake signon when served a
redirect. This did mean some of the tests are quite manual.

In putting this together I discovered that the permission tests didn't
work correctly so I fixed the code so it could be tested. I also
discovered that tests were affected by the Rails cache of the temporary
so I configured that to be a null store.