Relax constraint on rubocop-govuk dependency
In #261 (specifically this commit [1]), most of the dependency
constraints were relaxed to use pessimistic constraints [2]. For some
reason this gem was pinned at 4.7.0, but there's no explanation in the
commit note or the PR description.

Since Rubocop seems to run OK with the latest v4 of rubocop-govuk, I
think it makes sense to relax this constraint. This should mean we don't
keep getting Dependabot PRs for every patch version of rubocop-govuk.

[1]: 0161fdf616508f114951c60bebba91dab685454b
[2]: https://thoughtbot.com/blog/rubys-pessimistic-operator