Enable PKCE extension in GDS OmniAuth Strategy
https://trello.com/c/59EBweBx

In https://github.com/alphagov/signon/pull/2312 we enabled the OAuth2
PKCE extension[1] in Signon.

In this commit we update our GDS OAuth2 OmniAuth Strategy to make use of
the PKCE extension. This means that any of our apps using this Gem will
benefit from the additional protection offered by the PKCE extension.

[1]: https://datatracker.ietf.org/doc/html/rfc7636