Fix deprecation warning
The presence of this default value for @@cache was causing the
following deprecation warning when the gem was required in other apps:

    DEPRECATION WARNING: Support for
    `config.active_support.cache_format_version = 6.1` has been
    deprecated and will be removed in Rails 7.2.

This is because a newer cache format has been introduced[1].

This configuration option is rarely used across alphagov[2]. When it
is used it is only to set the value to `Rails.cache`. The railtie
`lib/gds-sso/railte.rb` already does this when the gem is used in
Rails applications.

We will need to bump the major version of this gem however, since this
change could theoretically cause a change in behaviour.

[1] https://guides.rubyonrails.org/upgrading_ruby_on_rails.html#new-activesupport-cache-serialization-format
[2]
https://github.com/search?q=org%3Aalphagov+%22config.cache+%3D+%22&type=code