Pin version of rubocop-govuk
For GOV.UK Ruby projects that don't have a Gemfile.lock file (usually
gems) we pin the exact version of rubocop-govuk so that there is no
discrepancy about which version of rubocop-govuk is used.

Without this you can fall into a frustrating situation where a local
dev's version of rubocop-govuk is out of sync with the version used by
CI and linting errors are not consistent between them.

This reverses: https://github.com/alphagov/gds-sso/pull/280