Encapsulate authenticate user code to a method
This moves the rather basic code for authenticating a user into a
standalone method. The motivation for this is so that this check can be
used without logic repetition. The use case I have intended is to use
this code as part of Rack middleware so that we can authenticate a user
before running other middleware (Rack::Attack rate limiting middleware).

I made a couple of key implementation decisions here:

1. I chose to add it as a method on GDS::SSO rather than creating a
   dedicated class like AuthoriseUser as it was such trivial code and
   left the door open for further similar utility methods
2. I went for the slightly unconventional approach of adding a bang
   method without an equivalent without a bang, this seemed appropriate
   as it leaves the door open for an `authenticate_user` method that
   doesn't trigger a failure app if we want to check if a user is
   authenticated without needing them to be authenticated.

This code could potentially be also used in the ControllerMethods
module, I didn't do this with this change because it doesn't quite map
correctly with the logic in the controller and looked like it'd be a bit
messy.