Add an api_request_matcher option
This option is provided as a mechanism to determine whether a request is
for an API and thus bearer token authentication based on a lambda that
is given a rack request object.

The motivation for adding this is from working with a monolithic
application that has both API and regular authenticated endpoints.
Adding this allows the app to return more appropriate errors, should
someone try access the API without a bearer token, and more accurate
dev/test environment as this ensures the bearer token strategy is used.

I thought the best way to do this matcher would be to have a lambda
function to provide client flexibility rather than set a regex that
matches a path. I thought for the easiest interface it'd be best to pass
in a Rack::Request object rather than the env hash - this does mean that
each call of this will have the overhead of the extra object allocation
for this but this should be negligible in the grand scheme of object
allocations per request.

I added the Rack gem as a dependency to provide this but this shouldn't
add any dependencies to an app using this as Rack is already a
sub-dependency of existing dependencies. Given that an app that uses
this likely has a large number of Rack dependencies and that we barely
utilise Rack's API I thought it would be least painful to lack a version
constraint.