Consider Config.api_only in checking api_call?
The config option of whether this Gem is in api_only mode should be
considered when api_call? is checked for completeness. Otherwise there's
a need to remember to consider api_only and api_call? whenever we make
decisions on if we're dealing with an API request.

This will likely impact apps that are already API only and in dev and
test environments may have been inadvertently using the mock_gds_sso
strategy rather than the mock bearer token one.