Return different responses for API failures
This updates the failure app to reflect the change in api_call?
requirements. In order to do this I have switched from making queries
against the rack env directly and instead stored data in that env during
the initialisation of Warden. The reason I have used this strategy is
that Warden modifies the rack env when it starts the failure app [1] and
thus API request matching can fail.

Changing to storing the data in the rack env does offer an advantage
that it can be effectively used as a cache so the same checks don't need
to be made multiple times.

[1]: https://github.com/wardencommunity/warden/blob/67f5ba6baaa7564ec79afef02cf3a4d0f7d312e5/lib/warden/manager.rb#L138-L140