Add system tests for mock strategies
These have been added in as I found that the API mock strategy only
works if there is a bearer token provided and thus is unlikely to be
activated as expected in a number of application tests where instead the
session based login is likely used.

This may not have been previously a problem but now there's a risk of
having no strategy as we've made the API call check stricter so for dev
mode you can find yourself unable to authenticate without providing a bearer
token.