Use own strategy for bearer token auth
This resolves an issue where apps could have no valid auth strategy
because the session based one returns false to valid? yet the bearer
token one is also not valid as a request doesn't have a bearer token.

This issue bites us specifically in dev/test environments where the
norm is to send requests without auth and expect the mock auth to kick
in. As this barely happens I expect that there is only very limited use
of the MockBearerToken strategy in tests. For this reason I've flagged
this as a breaking change, even though it's probably more of a bug fix.

Creating our own bearer strategy offered us an advantage of caching the
user we've looked up for the duration of the strategy. In the previous
iteration we'd be querying the DB for the user 6 times during an
individual auth cycle [1].

[1]: https://github.com/opperator/warden-oauth2/blob/master/lib/warden/oauth2/strategies/token.rb#L11-L19