Use own strategy for bearer token auth
This resolves an issue where apps could have no valid auth strategy
because the session based one returns false to valid? yet the bearer
token one is also not valid as a request doesn't have a bearer token.

This issue bites us specifically in dev/test environments where the
norm is to send requests without auth and expect the mock auth to kick
in. As this barely happens I expect that there is only very limited use
of the MockBearerToken strategy in tests. For this reason I've flagged
this as a breaking change, even though it's probably more of a bug fix.

Creating our own bearer strategy offered us an advantage of caching the
user we've looked up for the duration of the strategy. In the previous
iteration we'd be querying the DB for the user 6 times during an
individual auth cycle [1].

There were a couple of implementation decisions I made that are best
recorded:

1) changing the approach to find a bearer token from using
   env["HTTP_AUTHORIZATION"] to using Rack::Auth::AbstractRequest::AUTHORIZATION_KEYS

   This was done to not lose consistency with the strategy that was
   being replaced [2] which checks more headers than just
   HTTP_AUTHORIZATION. I don't think it's likely that there are other
   uses but it seemed unnecessary to break backwards compatibility here.

2) checking for a token_string to be non-empty in BearerToken.locate,
   this has been done to avoid querying away for a user when it will
   fail (and if things are worse if it succeeded). This check has to be
   done in the BearerToken class as the MockBearerToken needs to operate
   with an empty token

3) using a convoluted define? based memo approach on @token in
   :gds_bearer_token strategy. As the locate method executes a DB query
   that returns nil on no user found, this more elaborate memo approach
   is used so we don't query again and again when a request won't result
   in a found user.

[1]: https://github.com/opperator/warden-oauth2/blob/master/lib/warden/oauth2/strategies/token.rb#L11-L19
[2]: https://github.com/opperator/warden-oauth2/blob/9a229d35bd6804b74d515ca2744c9f3d5173566c/lib/warden/oauth2/strategies/bearer.rb#L15-L22