Return GDS::SSO.test_user without modifications for MockBearerToken
This removes an oddity where the config value of
additional_mock_permissions can be a bit overreaching as it was
previously overriding any permissions set on GDS::SSO.test_user.

This made it very difficult to test APIs where permissions were missing
as any permissions set on an existing user were overwritten when the
user was located.

I've flagged this as a breaking change but I expect the consequences to
be rather small as I don't imagine users expected this behaviour and
thus had workflows based around it and that MockBearerToken has likely
been the subject of rather low usage [1].

[1]: https://github.com/alphagov/gds-sso/pull/323