Consider GDS SSO's own API in matching API requests
This fixes a bug where using the new `api_request_matcher` option
blocked the use of Signon's own API as that failed the matcher.

The approach I've taken to resolve this is to add an additional matcher
that will match for Signon's own API. I thought this was better than
expecting clients to consider Signon's API as it is an implementation
detail you don't need to consider as someone setting up Signon in your
app.

As it's a controller that could be mounted at a different location I
thought it'd be prudent to set it up as something configurable even
though I'd expect it's very unlikely a user changes it.