Replace system test with request tests
This replaces the system tests for this app with request tests, starting
a process to introduce the request test concept centrally to this repo. The
motivation for this is:

- We were often breaking conventions of system tests in the system test
  by testing HTTP details such as headers and status codes
- The tests weren't really system tests as they didn't follow user
  interactions more prescribed HTTP interactions
- The system test was becoming the one place to do any integration tests
  making it rather bloated and something you had to somewhat fudge new
  test into

I've created a helper for requests that is intended to be used by other
request tests for this app which will replace the controller tests.

I didn't replicate the PKCE tests as they seem like they should be
associated with the auth controller and I'll re-introduce them there.

I chose to use ActiveSupport for time management rather than Timecop as
we already have the Rails dependency. I plan to remove Timecop entirely
later in this PR.