Reload routes between request tests to fix flakey tests
The random order of tests can cause the request tests to fail. This was
caused because the GDS::SSO::Config.api_only value affects which routes
are loaded, and as routes are loaded lazily if the tests start with one
that has mocked that value, then routes are missing.

The approach I found that fixed this was having to do a conditional
reload. If we just reloaded the routes every time (i.e. with
Rails.application.reload_routes!) then this would prevent routes loading
properly on the first request test that used them.

These failures only seem to occur in Rails versions newer than 7 where
it looks like lazy route loading was introduced [1]

[1]: https://github.com/rails/rails/pull/52353