Add request specs for AuthenticationController
This adds tests for a controller that wasn't previously tested in full.
It includes migrated tests from the previous
system/authentication_and_authorisation_spec.rb for the PKCE
configuration. I have moved them here as it seemed more appropriate to
test them near the endpoint, even if that configuration isn't explicit
in the controller.

Motivation for this is to have a more complete picture of integration
tests for HTTP endpoints.