Don't store very long return_to paths
We sometimes get `ActionDispatch::Cookies::CookieOverflow` errors and we
know that we can trigger these by trying to access a really long path
before being redirected to a sign in page. In [Signon][], we found that
if a full path (path plus query) was longer than 2048 bytes it would
cause this error, so we're setting the same limit here. If a `return_to`
path would exceed this limit, we skip storing it in the cookie. It's
unlikely any genuine path would come close to this, but if it does, we
think it's okay just to be redirected to root in these cases (by
`AuthenticationsController#callback`)

The motivation for exploring this was session cookie overflow errors in
Authenticating Proxy. We don't think these specific errors were caused
by really long paths, but this change provides a useful guard
nonetheless. We haven't seen the Authenticating Proxy session cookie
overflows for a while, but if and when we do we should look for other
places where we might store variable length values in the session cookie

[Signon]: https://github.com/alphagov/signon/pull/3963

Co-authored-by: Ynda Jas <yndajas@gmail.com>