Add User#anonymous_user_id and test it
Note that the tests match the behaviour in
https://github.com/alphagov/signon/pull/3992, including that given the
same secret and UID, we get the same value of anonymous_user_id.

I've also added an assertion into lint/user_spec.rb, which will check
that we generate consistent anonymous_user_id values for all User
classes which include the module.

I've deviated ever so slightly from the implementation in signon, in
that I've used `JSON.dump(stuff)` instead of `stuff.to_json`. This is
just because this Gem might be used in places without rails /
activesupport, so I'm not certain the .to_json method will always be
there.